gigamegapush

+++ Scripts/GameVariable.cs
@@ -7,7 +7,7 @@ public class GameVariable : MonoBehaviour
     public bool isGameRunning;
     public bool openShop = false;
     //TODO IMPLEMENTA
-    public byte difficolta; // 0 = facile, 1 = normale, 2 = difficile, 3 = hardcore
+    public int difficolta; // 0 = facile, 1 = normale, 2 = difficile, 3 = hardcore
     public Transform player;
     public int monete;
     public int moltiplicatoreMonete = 1;
